Migrate question creation page to a modal
